datapath: quick flow lookup for source and destination MAC addresses
This patch implements a least-possible flow extraction
and lookup for L2 source and destination MAC address,
which is activated when all the flows in the datapath
involve in only source and destination MAC addresses
(i.e, the other fields are wildcarded).
This code improves forwarding rate between 10 Gbps
NICs by 12 % for 64 byte frames.